# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Remolonas Engine API",
        "description": "API para la gestión de suscripciones de cajas orgánicas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar historial de pedidos del usuario",
                "operationId": "5894f7ae19272fefaf73e3a6f9d9782d",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Ver detalles de un pedido",
                "operationId": "f4375fed7d76265a3325423e2534e437",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/plans": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Listar todos los planes activos",
                "operationId": "60e20f16af8b1fba04bb0e5138c87e81",
                "responses": {
                    "200": {
                        "description": "Lista de planes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plan"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/plans/{plan}": {
            "get": {
                "tags": [
                    "Plans"
                ],
                "summary": "Ver detalles de un plan",
                "operationId": "809325f5d80ccbcadf15d7e66044cabc",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del plan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plan"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar todos los productos",
                "operationId": "c821c7f9e41e08594ad4eae5a99e812e",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Ver detalles de un producto",
                "operationId": "8cacdd61a07ddbbac50707e7a6b41036",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Listar suscripciones del usuario",
                "operationId": "4e0263830d492a9c5c0a615873461385",
                "responses": {
                    "200": {
                        "description": "Lista de suscripciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Subscription"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Crear una nueva suscripción",
                "operationId": "92d1ae2f0d00a5bb6cc5c3f5f5dafd68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "type": "integer"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "weekly",
                                            "monthly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Suscripción creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/subscriptions/{subscription}": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Ver detalles de una suscripción",
                "operationId": "df2c81af1896e7c97b55e560454a1e7d",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la suscripción",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Cancelar/Eliminar una suscripción",
                "operationId": "e48295d3cb8d3da3e5b1b79c52b88b82",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Suscripción eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Actualizar estado o frecuencia de una suscripción",
                "operationId": "c8d456f9428192207e891b6e527b4cbe",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "paused",
                                            "cancelled"
                                        ]
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "weekly",
                                            "monthly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Suscripción actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "subscription_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string"
                    },
                    "delivery_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Plan": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "available_quantity": {
                        "type": "integer"
                    },
                    "image_url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Subscription": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string"
                    },
                    "frequency": {
                        "type": "string"
                    },
                    "next_delivery_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "plan": {
                        "$ref": "#/components/schemas/Plan"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Plans",
            "description": "Plans"
        },
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "Subscriptions",
            "description": "Subscriptions"
        }
    ]
}